pyproject.toml
[tool.poetry]
name = "bloom"
version = "0.1.0"
description = ""
authors = ["madEffort <madeffort0903@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
django = "^5.0.4"
python-decouple = "^3.8"
django-widget-tweaks = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
